fix(db): Corrected misspelling of 'deleted' in billing_entities model
Fixed a typo in the 'billing_entities' model. The property 'deteled_at' was misspelled and has been updated to 'deleted_at'. This typo was impacting further code functionality as it was causing dependency issues and errors during runtime. Thus, this commit will help in removing such bugs and maintaining code consistency.