feat(db): Add scopes for invoice and billing entities in db
Added the ability to filter invoices and billing_entities by their deletion status.
Added three scopes, `NOT_DELETED`, `DELETED` for `INVOICE` and `NOT_DELETED` for `BILLING_ENTITY`. This gives the functionality to further filter `invoices` and `billing_entities` based on whether they are deleted or not.
Modified `initScopes` function to include the newly added scopes. This helps in better segregation of data and improves the querying capabilities of the application.